Oops.. explicitly removing 'Copyleft 🄯' (conflict with Public Domain)
On a quick review, just realized that the 'Copyleft' copy & pasted over
and over again obviously confict with Public Domain. Even the idea of
allowing license with either Public Domain OR BSD0 was to allow
usage on places that do not recognize public domain (commercial
or not, so letting this 'Copyleft 🄯' was literally an mistake.
(maybe actually worse than if was a generic copyright)

The 'Copyleft 🄯' was simply replaced with 'Author: '. Also betwen
the author name and the e-mail is the (Etica.AI) just to avoid
problems with whatever organization I could be working/colaborating